# Add hash method to metadata to speed up comparing it in a hash key