@repeatedly 2019-08-16
`timekey.object_id` is safe for existing use cases?
metadata is used in buffer for hash key.